feat: add uint = unsigned int to most files

+++ src/lib/ite.h
@@ -2,6 +2,7 @@
 #include <string>
 #include <vector>
 #include "CImg.h"
+#include "core/utils.h"
 
 using namespace cimg_library;
 
@@ -214,7 +215,7 @@ namespace ite
         /** @brief The kernel size for median denoising (default 3). */
         int median_kernel_size = 3;
         /** @brief The threshold for median denoising (default 0). */
-        float median_threshold = 0;
+        int median_threshold = 0;
         /** @brief Maximum window size for adaptive median filter (must be odd, default 7). */
         int adaptive_median_max_window = 7;
